TIP-874: move TimestampableInterface to Versioning
