+++ src/Akeneo/Pim/Enrichment/Bundle/spec/EventSubscriber/TimestampableSubscriberSpec.php
@@ -5,7 +5,7 @@ namespace spec\Akeneo\Pim\Enrichment\Bundle\EventSubscriber;
 use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 use PhpSpec\ObjectBehavior;
-use Akeneo\Pim\Enrichment\Component\Product\Model\TimestampableInterface;
+use Akeneo\Tool\Component\Versioning\Model\TimestampableInterface;
 use Prophecy\Argument;
 
 class TimestampableSubscriberSpec extends ObjectBehavior
