+++ src/Akeneo/Pim/Structure/Component/Model/AssociationTypeInterface.php
@@ -5,7 +5,7 @@ namespace Akeneo\Pim\Structure\Component\Model;
 use Akeneo\Tool\Component\Localization\Model\TranslatableInterface;
 use Akeneo\Tool\Component\StorageUtils\Model\ReferableInterface;
 use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\TimestampableInterface;
+use Akeneo\Tool\Component\Versioning\Model\TimestampableInterface;
 
 /**
  * Association type
